Add better error handling and add comments

+++ pkg/realtime_api/common.go
@@ -9,7 +9,7 @@ type incommingMessage struct {
 
 type outgoingMessage struct {
     Type        string          `json:"type"`
-    Result      uint8           `json:"result,omitempty"`
+    Result      uint16          `json:"result,omitempty"`
     Obj         []interface{}   `json:"obj,omitempty"`
     Id          uint64          `json:"id"`
 }
